added db.php to separtate concerns and allow for testing. Fixed typos in code on index.html
